### 2025-05-26 #8274 @editorialbot
[PRE REVIEW]: pyElli: An open source ellipsometry analysis tool for FAIR data
**Submitting author:** @MarJMue (<a href="http://orcid.org/0009-0005-2187-0122">Marius J. Müller</a>)
**Repository:** https://github.com/PyEllips/pyElli
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 0.22.3
