+++ paper/paper.bib
@@ -209,6 +209,7 @@
   eprint={1603.02720},
   archivePrefix={arXiv},
   primaryClass={physics.comp-ph}
+  doi={10.48550/arXiv.1603.02720}
 }
 
 @article{Scheidgen2023,
@@ -390,7 +391,7 @@
   pages = {108256},
   year = {2022},
   issn = {0010-4655},
-  doi = {https://doi.org/10.1016/j.cpc.2021.108256},
+  doi = {10.1016/j.cpc.2021.108256},
   url = {https://www.sciencedirect.com/science/article/pii/S0010465521003684},
   author = {Mélanie M. Bay and Silvia Vignolini and Kevin Vynck},
   keywords = {Multilayers, Anisotropic optical materials, Optical modelling, Photonic crystals, Cholesterics, Surface phonon polaritons},
