### 2025-05-26 bed5a14
add Python Benchmark with pytest-benchmark (pytest) benchmark result for e753c341a5cb23b6036403e42fcea4802fd11081
