### 2025-05-26 c0b0a47
Badge formatting
