### 2025-05-26 71b4465
add Python Benchmark with pytest-benchmark (pytest) benchmark result for c0b0a47f8572148705ba6f6169817f3281a6bb09
