### 2025-05-27 922b253
Remove doi for Byrnes
