+++ paper/paper.bib
@@ -209,7 +209,6 @@
   eprint={1603.02720},
   archivePrefix={arXiv},
   primaryClass={physics.comp-ph}
-  doi={10.48550/arXiv.1603.02720}
 }
 
 @article{Scheidgen2023,
