+++ .github/workflows/build-docker.yml
@@ -26,7 +26,7 @@ jobs:
           images: domna/pyelli
       
       - name: Build and push Docker image
-        uses: docker/build-push-action@1dc73863535b631f98b2378be8619f83b136f4a0
+        uses: docker/build-push-action@263435318d21b8e681c14492fe198d362a7d2c83
         with:
           context: .
           push: true
