### 2025-06-02 feeae10
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 49916db6b824e517a57003f5c2ca9c902362bc92
