### 2025-06-10 1ad0fef
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 7fa2315d2ef39467456f64f31a27abce720a336c
