+++ docs/requirements.txt
@@ -1808,9 +1808,9 @@ uncertainties==3.2.2 \
     # via
     #   -r requirements/dev-requirements.txt
     #   lmfit
-urllib3==2.3.0 \
-    --hash=sha256:1cee9ad369867bfdbbb48b7dd50374c0967a0bb7710050facf0dd6911440e3df \
-    --hash=sha256:f8c5449b3cf0861679ce7e0503c7b44b5ec981bec0d1d3795a07f1ba96f0204d
+urllib3==2.5.0 \
+    --hash=sha256:3fc47733c7e419d4bc3f6b3dc2b4f890bb743906a30d56ba4a5bfa4bbff92760 \
+    --hash=sha256:e6b01673c0fa6a13e374b50871808eb3bf7046c4b125b216f6bf1cc604cff0dc
     # via requests
 uv==0.6.5 \
     --hash=sha256:15dae245979add192c4845947da1a9141f95c19403d1c0d75019182e6882e7d4 \
