### 2025-06-19 9a1520c
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 0c1e2542933bd6f84b387cb33dcbfba80f2ae49c
