### 2025-06-21 #8274 @zhubonan
Hi @phoebe-p and @Mil152, would you be able to help review this submission, which I think is in your area of expertise?

The reviewing process of JOSS is unique as it happens on GitHub. The process is transparent and interaction between the reviewers and the author are encouraged. You can learn about the reviewing guideline here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html.
