### 2025-06-23 578a651
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 38443ec20fa6b07fc5a299f20f03d92ec6239cd9
