### 2025-06-23 61638f8
Fix for wollam data without dpolE
