### 2025-06-23 fe38105
Use union typing to support older py versions
