### 2025-06-25 5d2a398
Fix `as_index()` and `as_dielectric()`
