### 2025-06-25 fdbfded
Add test to check if values stay the same after dispersion conversion
