### 2025-06-25 bf7cdeb
Remove `as_dielectric()`
