+++ tests/test_dispersions.py
@@ -244,20 +244,29 @@ def test_deepcopy_of_index_dispersion():
     deepcopy(sell.as_index())
 
 
+def test_index_dispersion_conversion():
+    sell = Sellmeier()
+    sell.add(1, 1)
+
+    lbda = np.linspace(300, 900, 100)
+
+    assert_array_equal(
+        sell.as_index().get_dielectric(lbda),
+        sell.get_dielectric(lbda),
+    )
+
+    assert_array_equal(
+        sell.as_index().add(2, 2).get_dielectric(lbda), sell.get_dielectric(lbda)
+    )
+
+
 def test_dispersion_conversion_to_other_type():
     lbda = np.linspace(300, 900, 100)
     RII = elli.db.RII()
 
-    disp = RII.get_dispersion("SCHOTT-BK", "N-BK7")
+    disp = RII.get_dispersion("AgCl", "Tilton")
 
     assert_array_equal(
-        disp.as_dielectric().get_dielectric(lbda),
+        disp.as_index().get_dielectric(lbda),
         disp.get_dielectric(lbda),
-    )
-
-    disp2 = RII.get_dispersion("AgCl", "Tilton")
-
-    assert_array_equal(
-        disp2.as_index().get_dielectric(lbda),
-        disp2.get_dielectric(lbda),
     )
