+++ src/elli/dispersions/base_dispersion.py
@@ -2,7 +2,6 @@
 """Abstract base class and utility classes for pyElli dispersion"""
 
 from abc import ABC, abstractmethod
-from copy import deepcopy
 from typing import List, Optional, Union
 
 import numpy as np
