### 2025-06-27 21dc4bd
Fix for wollam data without dpolE (#214)
* Fix for wollam data without dpolE

* Use union typing to support older py versions

* Fix `as_index()` and `as_dielectric()`

* Add test to check if values stay the same after dispersion conversion

* Remove `as_dielectric()`

* Fix tests

* Set rep and single params

* Remove unecessary import

* Update changelog

---------

Co-authored-by: MarJMue <49639740+MarJMue@users.noreply.github.com>
