### 2025-06-27 3608b68
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 21dc4bd3ae139830789c410fcbd7d1d7596c7aa9
