### 2025-07-31 #8503 @phoebe-p
Minor comment on the installation instructions: I believe this issue is specific to (the default zsh terminal on) MacOS, but running `pip install pyElli[fitting]` doesn't work - commands with square brackets need to have quotation marks around them. So `pip install "pyElli[fitting]"` does work. I don't know if that causes issues on other operating systems/terminals though.
