### 2025-07-31 #8503 @phoebe-p
I think the paper overall gives a good overview. I have some minor comments:

- The 'FAIR' acronym isn't defined anywhere in the text
- What is the word 'respective' doing on line 13 (it seems unnecessary to me)?
- 'Competitive systems' should be 'competing systems'
- A minor point, but in the 'similar software' section, RayFlare also uses a modified, vectorized (over wavelengths) version of Byrnes' 2x2 TMM code
- Line 132: redundant degree symbol + the word 'degree'
