### 2025-07-31 #8503 @Mil152
A standard `git clone` does not fetch the contents of submodules. For the package to be installed correctly from source, the installation instructions must be updated to include the step for initializing submodules.

For users who have already cloned the repository:

```bash
git submodule update --init --recursive
```

For a fresh clone, the command should be:

```bash
git clone --recurse-submodules https://github.com/PyEllips/pyElli.git
```

I recommend adding this information to the installation section of the `README.md` file.
