### 2025-08-04 23ee2c0
add Python Benchmark with pytest-benchmark (pytest) benchmark result for d84c5c970ddbde013f45809c46a6e286aed90266
