### 2025-08-06 c9d40b8
Update install instructions to use recurse submodule and pip install with quotes
