### 2025-08-06 PyEllips/pyElli#218 @domna
Run pytorch test only if dependency is installed
From https://github.com/openjournals/joss-reviews/issues/8503#issuecomment-3141235188

> [...] The test `tests/test_TiO2.py::TestTiO2::test_solver4x4_expm_torch` failed with an `ImportError: PyTorch is not installed`. The core test suite should not fail if an optional dependency like PyTorch is not installed. Tests for optional features should be skipped automatically. Please modify this test to use `pytest.mark.skipif`. [...]
