### 2025-08-06 PyEllips/pyElli#219 @domna
Improve examples
From https://github.com/openjournals/joss-reviews/issues/8503#issuecomment-3141365102

> For examples in the package, The `Basic Usage.ipynb` notebook is excellent. It contains clear descriptive text in markdown cells that effectively explains the purpose of the code and guides the user through the workflow. However, the other notebooks in the directory consist primarily of code cells and lack this explanatory text. This makes it a bit difficult for a user to follow the logic or understand the context behind each step. Update the other example notebooks to include descriptive markdown cells, using `Basic Usage.ipynb` as a model. This will significantly improve their value as a learning resource and make the package more accessible to the community.
