### 2025-08-07 1540a03
add Python Benchmark with pytest-benchmark (pytest) benchmark result for d37e33d450b8e7ba2b094de5dbf92ed10efa890d
