### 2025-08-07 1c30753
add Python Benchmark with pytest-benchmark (pytest) benchmark result for b8d3ba6040d92d4ddac633ca7dbb9028dfeebae6
