### 2025-08-08 d2127c8
Bump uv from 0.6.5 to 0.8.6 in /docs (#221)
