+++ docs/requirements.txt
@@ -1812,25 +1812,26 @@ urllib3==2.5.0 \
     --hash=sha256:3fc47733c7e419d4bc3f6b3dc2b4f890bb743906a30d56ba4a5bfa4bbff92760 \
     --hash=sha256:e6b01673c0fa6a13e374b50871808eb3bf7046c4b125b216f6bf1cc604cff0dc
     # via requests
-uv==0.6.5 \
-    --hash=sha256:15dae245979add192c4845947da1a9141f95c19403d1c0d75019182e6882e7d4 \
-    --hash=sha256:23aa8e8ca7795f54f6cf0f0fbd0aaf7b26bf4aae42f8c10643bcba6d42485a3f \
-    --hash=sha256:26a90e69d6438de2ec03ab452cc48d1cb375249c6b6980f4ed177f324a5ad8b3 \
-    --hash=sha256:43847ef95d56c239de940339e5cfc2ade58249005e8ab97244fdb69fb9761572 \
-    --hash=sha256:442639a874f6bb6864279f099c97739287d7e244bc25d0f791345cc69f46c940 \
-    --hash=sha256:5323e9219a519c6553111820a8c54588d426380404a208b23cf4c3265bc87ec6 \
-    --hash=sha256:6210fe6ef6a0ae3dc618611fcc8ada4e620fea5172fb8a9c50d3a59b6915b023 \
-    --hash=sha256:6e82de1cb6a116f7736de9542430d78c210d152c80723db8beffc14e5b4e4b40 \
-    --hash=sha256:6f3b3f21b2385545f499f6cc21f44eac3bbb0f6cb98fbf9c6d3e58db186c8a41 \
-    --hash=sha256:70ad4cc5f2b636edbeeebb3aee0a7daa66b17457038088be870ac7adc5a9842d \
-    --hash=sha256:776500595ff7cda1ffa5a76dd3ff9de3819f1e26c493938cbdc20c1ab766b2eb \
-    --hash=sha256:9b56fa88951fab3bc7164255d844de9ad048e6a04a95f1c2774637e06889efe6 \
-    --hash=sha256:a481254f63240023239ecec80cd690dec05875e248eb4b9d7f66957b017798b1 \
-    --hash=sha256:b5445a509f500bbf18faba4e7cf5cc9763617c335d58afaa5f3e5a6e388dd4ee \
-    --hash=sha256:c5676fc7cdd088e2c3342593c1d2dc379bf86a83301af7b0dfe8d45801a50d85 \
-    --hash=sha256:c5683bccfc2b92cdc2f91e2904baa8ee2b5893b33ac8acac25e702ce7d3e5415 \
-    --hash=sha256:d47b4adffcdbe30bd678c7670e63c671b8b34a667898501e588f2e7cbce34656 \
-    --hash=sha256:ef0b2f810d87aa9bbad15c3ab113e555871f14c9cd8ad2205338fb0358aaf52d
+uv==0.8.6 \
+    --hash=sha256:0b524de39f317bd8733c38cf100b6f8091d44e06b23f7752523ad1ad1454ede3 \
+    --hash=sha256:132e73f1e9fe05edc6c06c00416f7c721c48298786fd7293be6c584793170bbc \
+    --hash=sha256:1913f5627c57076c88dd38b0173bdb006ae9b8dbd92b1798a1acc9d744c1a7cc \
+    --hash=sha256:2ac28509db2e52613a59264bdb150d13274ed13e5b305f7e274da8cd83033985 \
+    --hash=sha256:37227aaf1e41c7eda3d7f0028e747a2a2eed3f3506b0adc121a4366e8281115b \
[…]
