### 2025-08-08 2f6c586
add Python Benchmark with pytest-benchmark (pytest) benchmark result for d2127c8a8862979a1a77bcb60312b4da721daaa9
