### 2025-08-11 0af1c28
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 922cf777535b3ca7e9e0e7a2ea58df5420fce643
