### 2025-08-11 85b6014
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 4f4680ed8887be46730907265fe5db0b6bdb5ec7
