### 2025-08-12 221cf6d
add Python Benchmark with pytest-benchmark (pytest) benchmark result for c9463420873d30adcc309d8d9884d76764f262c4
