+++ .github/workflows/pytest.yml
@@ -18,7 +18,7 @@ jobs:
         python_version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
 
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v5
         with:
           submodules: recursive
       - name: Set up Python ${{ matrix.python_version }}
