+++ .github/workflows/benchmark.yml
@@ -14,7 +14,7 @@ jobs:
     name: Run pytest-benchmark benchmark example
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v5
         with:
           submodules: recursive
       - uses: actions/setup-python@v5
