### 2025-08-18 def2f8d
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 3169c727084b382713ccd4f574534ce1c93a5444
