### 2025-09-01 0f4a4c0
Woollam: Fix importer regex for negative deltas
