### 2025-09-01 abf1f2d
Woollam: read first data line of file
