+++ tests/test_wollam.py
@@ -15,9 +15,9 @@ def test_reading_of_psi_delta_woollam(datadir):
     )
     data_cease = elli.read_woollam_psi_delta(datadir / "complete_ease_example.dat")
 
-    assert data_wvase.shape == (542, 2)
-    assert data_wvase_wo_dpolE.shape == (542, 2)
-    assert data_cease.shape == (3263, 2)
+    assert data_wvase.shape == (543, 2)
+    assert data_wvase_wo_dpolE.shape == (543, 2)
+    assert data_cease.shape == (3264, 2)
 
 
 # pylint: disable=redefined-outer-name
@@ -26,8 +26,8 @@ def test_reading_and_conv_to_woollam(datadir):
     data_wvase = elli.read_woollam_rho(datadir / "wvase_example.dat")
     data_cease = elli.read_woollam_rho(datadir / "complete_ease_example.dat")
 
-    assert data_wvase.shape == (542,)
-    assert data_cease.shape == (3263,)
+    assert data_wvase.shape == (543,)
+    assert data_cease.shape == (3264,)
 
 
 # pylint: disable=redefined-outer-name
