### 2025-09-07 198e5ce
Add wavelength boundary check
