+++ .github/workflows/python-publish.yml
@@ -12,7 +12,7 @@ jobs:
       with:
         submodules: recursive
     - name: Set up Python
-      uses: actions/setup-python@v5
+      uses: actions/setup-python@v6
       with:
         python-version: "3.x"
     - name: Install pypa/build
