+++ .github/workflows/requirements.yml
@@ -14,7 +14,7 @@ jobs:
         with:
           submodules: recursive
       - name: Set up Python 3.13
-        uses: actions/setup-python@v5
+        uses: actions/setup-python@v6
         with:
           python-version: "3.13"
       - name: Install package (no deps)
