+++ .github/workflows/pytest.yml
@@ -22,7 +22,7 @@ jobs:
         with:
           submodules: recursive
       - name: Set up Python ${{ matrix.python_version }}
-        uses: actions/setup-python@v5
+        uses: actions/setup-python@v6
         with:
           python-version: ${{ matrix.python_version }}
       - name: Install module
