### 2025-09-16 e824bff
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 83050a0e74070df2666debc4d4d600708eba5c3f
