### 2025-09-16 36dd713
Fixes for Woollam importer (#226)
* Woollam: Fix importer regex for negative deltas
* Woollam: read first data line of file
* Add wavelength boundary check

---------

Co-authored-by: domna <github@schroedingerscat.org>
