+++ src/elli/importer/woollam.py
@@ -46,7 +46,7 @@ def _is_wvase_tabular(line: str) -> bool:
     Returns:
         bool: True if the line is in wvase tabular layout.
     """
-    return bool(re.search(r"^(\d*\.\d*\s+){5}\d*\.\d*$", line))
+    return bool(re.search(r"^(-?\d*\.\d*\s+){5}\d*\.\d*$", line))
 
 
 def _is_complete_ease_tabular(line: str) -> bool:
@@ -181,11 +181,9 @@ def read_woollam_psi_delta(fname: str) -> pd.DataFrame:
         line = fobj.readline()
         while line:
             if _is_wvase_tabular(line):
-                line_number = fobj.tell()
                 file_format = "wvase"
                 break
             if _is_complete_ease_tabular(line):
-                line_number = fobj.tell()
                 file_format = "complete_ease"
                 break
             if _is_tan_cos_format(line):
@@ -194,6 +192,7 @@ def read_woollam_psi_delta(fname: str) -> pd.DataFrame:
                     "Please try using wvase's psi/delta format."
                 )
             metadata.append(line)
+            line_number = fobj.tell()
             line = fobj.readline()
         fobj.seek(line_number)
 
