### 2025-09-16 707e325
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 36dd713dbbdcab3719a3dd7aa1b53759c569ab20
