### 2025-09-16 46e62bd
Use chardet in RII scripts
