### 2025-09-22 cabf8fa
Add windows pytest
