+++ .github/workflows/pytest.yml
@@ -10,6 +10,26 @@ on:
     branches: [master]
 
 jobs:
+  test_python_win:
+    runs-on: windows-latest
+
+    steps:
+      - uses: actions/checkout@v5
+        with:
+          submodules: recursive
+      - name: Set up Python 3.13
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.13"
+      - name: Install module
+        run: |
+          curl -LsSf https://astral.sh/uv/install.sh | sh
+          uv pip install --system ".[fitting,dev]"
+          uv pip install --system torch --index-url https://download.pytorch.org/whl/cpu
+      - name: Test with pytest
+        run: |
+          pytest --nbmake
+
   test_python:
     runs-on: ubuntu-latest
     strategy:
@@ -25,9 +45,14 @@ jobs:
         uses: actions/setup-python@v6
         with:
           python-version: ${{ matrix.python_version }}
+
+      - name: Install uv (prebuilt binary)
+        run: |
+          curl -LsSf https://astral.sh/uv/install.ps1 | powershell -Command -
+          echo "$env:USERPROFILE\.cargo\bin" | Out-File -FilePath $env:GITHUB_PATH -Encoding utf8 -Append
+
       - name: Install module
         run: |
-          curl -LsSf https://astral.sh/uv/install.sh | sh
           uv pip install --system ".[fitting,dev]"
           uv pip install --system torch --index-url https://download.pytorch.org/whl/cpu
       - name: Test with pytest
