### 2025-09-22 1de78ac
Fix installation of uv
