+++ .github/workflows/pytest.yml
@@ -22,10 +22,8 @@ jobs:
         with:
           python-version: "3.13"
 
-      - name: Install uv
-        run: |
-          curl -LsSf https://astral.sh/uv/install.ps1 | powershell -Command -
-          echo "$env:USERPROFILE\.cargo\bin" | Out-File -FilePath $env:GITHUB_PATH -Encoding utf8 -Append
+      - name: Install the latest version of uv
+        uses: astral-sh/setup-uv@v6
 
       - name: Install module
         run: |
