### 2025-09-22 ea6f183
Add dispersion with encoding errors
