### 2025-09-23 ea10616
Fix for rii encoding problems under windows + windows pytest (for 3.13 only) (#228)
* Use chardet in RII scripts

* Add windows pytest

* Fix installation of uv

* Us uv install action

* Add dispersion with encoding errors

* Cleanup diff

---------

Co-authored-by: MarJMue <49639740+MarJMue@users.noreply.github.com>
