### 2025-09-23 234a5fe
add Python Benchmark with pytest-benchmark (pytest) benchmark result for ea10616061fe41ea3c6be5b5b42f290279440b8a
