### 2025-09-29 08df806
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 8d8c2e69f0e93eded238146ae3e771847785a634
