### 2025-10-07 5dc70bc
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 6f91ec683102ec47b740114394080ea224069fdc
