+++ .github/workflows/pytest.yml
@@ -23,7 +23,7 @@ jobs:
           python-version: "3.13"
 
       - name: Install the latest version of uv
-        uses: astral-sh/setup-uv@v6
+        uses: astral-sh/setup-uv@v7
 
       - name: Install module
         run: |
