### 2025-10-13 369ecb2
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 3f320f7c8b28b70879056b06ccd7b73af9502fe9
