### 2025-10-14 b5c9fc0
Update gallery plots
