+++ examples/gallery/plot_interface_reflection.py
@@ -4,12 +4,21 @@ Interface reflection
 """
 
 # %%
-# Interface between two materials
-# -------------------------------
+# Reflection on interfaces between two materials
+# ==============================================
+#
 # Authors: O. Castany, C. Molinaro, M. Müller
 #
-# Interface between two materials n1/n2.
-# Calculations of the transmission and reflexion coefficients with varying incidence angle.
+# This notebook analyzes the optical behavior at the interface between two
+# isotropic materials with refractive indices n1 and n2.
+#
+# It computes the reflection and transmission coefficients as functions of
+# the incidence angle, for both s- and p-polarized light, using the
+# analytical Fresnel equations and comparing those to pyElli’s output.
+
+# %%
+# Import required libraries
+# -------------------------
 import elli
 import matplotlib.pyplot as plt
 import numpy as np
@@ -17,39 +26,61 @@ from scipy.constants import c, pi
 
 # %%
 # Structure definition
-# ------------------------
+# --------------------
+#
+# We define an optical system with two materials:
+#
+# -  Frontside medium (air) with refractive index n = 1.0
+# -  Back medium (glass) with n = 1.5
+#
+# We also prepare the vacuum wavevector k0 and a range of incidence angles
+# from 0° to 89°.
+
+# Define refractive indices for the two media
 n1 = 1
 n2 = 1.5
+
+# Create isotropic materials using pyElli's ConstantRefractiveIndex model
 front = elli.IsotropicMaterial(elli.ConstantRefractiveIndex(n1))
 back = elli.IsotropicMaterial(elli.ConstantRefractiveIndex(n2))
 
-# Structure
+# Define the optical structure: just a single interface with no internal layers
 s = elli.Structure(front, [], back)
 
-# Parameters for the calculation
+# Set wavelength and wavevector
 lbda = 1000
 k0 = 2 * pi / lbda
-Phi_list = np.linspace(0, 89, 90)  #  range for the incidence angles
+
+# Define a range of incidence angles from 0° to 89°
+Phi_list = np.linspace(0, 89, 90)
 
 # %%
 # Analytical calculation
 # ------------------------
+# We calculate the Fresnel reflection and transmission coefficients for s- and p-polarized light at an interface.
+
+# Convert incidence angles from degrees to radians
 Phi_i = np.deg2rad(Phi_list)
 
+# Snell's law (allowing complex values for total internal reflection)
 Phi_t = np.arcsin((n1 * np.sin(Phi_i) / n2).astype(complex))
+
 kz1 = n1 * k0 * np.cos(Phi_i)
 kz2 = n2 * k0 * np.cos(Phi_t)
+
+# Fresnel coefficients
 r_s = (kz1 - kz2) / (kz1 + kz2)
 t_s = 1 + r_s
 r_p = (kz1 * n2**2 - kz2 * n1**2) / (kz1 * n2**2 + kz2 * n1**2)
 t_p = np.cos(Phi_i) * (1 - r_p) / np.cos(Phi_t)
 
-# Reflection and transmission coefficients, polarisation s and p
+# Reflectance and Transmittance
 R_th_ss = abs(r_s) ** 2
 R_th_pp = abs(r_p) ** 2
 t2_th_ss = abs(t_s) ** 2
 t2_th_pp = abs(t_p) ** 2
-# The power transmission coefficient is T = Re(kz2/kz1) × |t|^2
+
+# The power transmission coefficient (correction for energy conservation) is T = Re(kz2/kz1) × |t|^2
 correction = np.real(kz2 / kz1)
[…]
