### 2025-10-14 a70eb7c
Fix wavelength slicing in gallery
