### 2025-10-15 d0469a2
Add changelog for version 0.22.5
