### 2025-10-15 27afcdc
add Python Benchmark with pytest-benchmark (pytest) benchmark result for d0469a241f8df087fd93d31e89cd2a103fd30116
