+++ docs/reference/api_documentation.md
@@ -1,0 +1,9 @@
+# API Documentation
+
+## Kramers-Kronig relations
+
+::: elli.kkr
+
+## Refractiveindex.info database
+
+::: elli.db.RII
