+++ docs/authors.rst
@@ -1,2 +1,0 @@
-.. _authors:
-.. mdinclude:: ../AUTHORS.md
