+++ docs/changelog.rst
@@ -1,2 +1,0 @@
-.. _changes:
-.. mdinclude:: ../CHANGELOG.md
