+++ docs/requirements.txt
@@ -1,1855 +1,0 @@
-# This file was autogenerated by uv via the following command:
-#    uv pip compile --extra=fitting --extra=dev --extra=docs --generate-hashes --output-file=docs/requirements.txt requirements/dev-requirements.txt pyproject.toml
-alabaster==1.0.0 \
-    --hash=sha256:c00dca57bca26fa62a6d7d0a9fcce65f3e026e9bfe33e9c538fd3fbb2144fd9e \
-    --hash=sha256:fc6786402dc3fcb2de3cabd5fe455a2db534b371124f1f21de8731783dec828b
-    # via sphinx
-anywidget==0.9.15 \
-    --hash=sha256:1891c11897aaf7cff8809f996413f618f97d786b6097f7e46266423969a726a0 \
-    --hash=sha256:fd7876332e47f380e0428f552f26b7227f5694d4e0a257bbc23354d9b9e9a73c
-    # via
-    #   -r requirements/dev-requirements.txt
-    #   pyelli (pyproject.toml)
-asteval==1.0.6 \
-    --hash=sha256:1aa8e7304b2e171a90d64dd269b648cacac4e46fe5de54ac0db24776c0c4a19f \
-    --hash=sha256:5e119ed306e39199fd99c881cea0e306b3f3807f050c9be79829fe274c6378dc
-    # via
-    #   -r requirements/dev-requirements.txt
-    #   lmfit
-asttokens==3.0.0 \
-    --hash=sha256:0dcd8baa8d62b0c1d118b399b2ddba3c4aff271d0d7a9e0d4c1681c79035bbc7 \
-    --hash=sha256:e3078351a059199dd5138cb1c706e6430c05eff2ff136af5eb4790f9d28932e2
-    # via
-    #   -r requirements/dev-requirements.txt
-    #   stack-data
-attrs==25.1.0 \
-    --hash=sha256:1c97078a80c814273a76b2a298a932eb681c87415c11dee0a6921de7f1b02c3e \
-    --hash=sha256:c75a69e28a550a7e93789579c22aa26b0f5b83b75dc4e08fe092980051e1090a
-    # via
-    #   -r requirements/dev-requirements.txt
-    #   jsonschema
-    #   referencing
-babel==2.17.0 \
-    --hash=sha256:0c54cffb19f690cdcc52a3b50bcbf71e07a808d1c80d549f2459b9d2cf0afb9d \
-    --hash=sha256:4d0b53093fdfb4b21c92b5213dba5a1b23885afa8383709427046b21c366e5f2
-    # via sphinx
-certifi==2025.1.31 \
-    --hash=sha256:3d5da6925056f6f18f119200434a4780a94263f10d1c21d032a6f6b2baa20651 \
-    --hash=sha256:ca78db4565a652026a4db2bcdf68f2fb589ea80d0be70e03929ed730746b84fe
-    # via requests
-cfgv==3.4.0 \
-    --hash=sha256:b7265b1f29fd3316bfcd2b330d63d024f2bfd8bcb8b0272f8e19a504856c48f9 \
-    --hash=sha256:e52591d4c5f5dead8e0f673fb16db7949d2cfb3f7da4582893288f0ded8fe560
-    # via
-    #   -r requirements/dev-requirements.txt
-    #   pre-commit
-chardet==5.2.0 \
[…]
