+++ docs/helpers.rst
@@ -1,8 +1,0 @@
-=================
-Helper functions
-=================
-Helper functions for generation rotation matrices and
-conversions for wavelengths and ellipsometric quantities.
-
-.. automodule:: elli.utils
-    :members:
