+++ docs/result.rst
@@ -1,8 +1,0 @@
-======
-Result
-======
-
-.. automodule:: elli.result
-    :members:
-    :undoc-members:
-    :show-inheritance:
