+++ docs/database.rst
@@ -1,6 +1,0 @@
-===================
-Dispersion Database
-===================
-
-.. autoclass:: elli.db.RII
-   :members:

[…]
