### 2025-10-15 5e5140e
Improve example descriptions (#229)
