+++ examples/gallery/plot_01_basic_usage.py
@@ -1,12 +1,12 @@
 """
-Basic usage
-===========
+Basic usage of pyElli
+=====================
 
 Basic usage of building a model and fitting it to measurement data of SiO2 on Si.
 
 """
 
-# %%
+# %% Import required libraries
 import elli
 from elli.fitting import ParamsHist, fit
 
@@ -26,7 +26,7 @@ from elli.fitting import ParamsHist, fit
 # This is because we're using literature values for Si,
 # which are only defined in this wavelength range.
 ANGLE = 70
-psi_delta = elli.read_nexus_psi_delta("SiO2onSi.ellips.nxs").loc[ANGLE][210:800]
+psi_delta = elli.read_nexus_psi_delta("SiO2onSi.ellips.nxs").loc[ANGLE].loc[210:800]
 
 # %%
 # Setting parameters
