### 2025-10-15 8014329
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 5e5140e03d119ede5d4c563d1714a73488c499a0
