### 2025-10-15 509da28
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 0d72bb7176d09eb105abad4c5fdd03d294d7452b
