### 2025-10-15 0d3fa32
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 5036ab6a7416f462bff7c6917ec3f88a942e001c
