### 2025-10-16 c78eb87
Revise TMMax citation in paper.bib
Updated the citation for TMMax software to include additional details such as eprint, archive prefix, primary class, and DOI.
