+++ paper/paper.bib
@@ -417,12 +417,15 @@
   year={2012}
 }
 
-@software{tmmax,
-  author = {Bahrem Serhat Danis, Esra Zayim},
-  title = {tmmax: transfer matrix method with jax},
-  version = {1.0.0},
-  url = {https://github.com/bahremsd/tmmax},
-  year = {2025}
+@misc{tmmax,
+  title        = {TMMax: High-performance modeling of multilayer thin-film structures using transfer matrix method with JAX},
+  author       = {Bahrem Serhat Danis and Esra Zayim},
+  year         = {2025},
+  eprint       = {2507.11341},
+  archivePrefix= {arXiv},
+  primaryClass = {physics.comp-ph},
+  doi          = {10.48550/arXiv.2507.11341},
+  url          = {https://doi.org/10.48550/arXiv.2507.11341}
 }
 
 @article{Langevin:24,
