### 2025-10-16 #8503 @bahremsd
Hi @MarJMue @domna , great work on this paper! 👏

I noticed that you kindly cited our work **TMMax** — thank you very much for that. Our paper has also been submitted to JOSS (see [JOSS review link](https://github.com/openjournals/joss-reviews/issues/9088)), and the latest version is now available on arXiv with a DOI. It would be great if you could update the citation to the arXiv version.

I’ve already opened a pull request for this change ([#235](https://github.com/PyEllips/pyElli/pull/235)) — feel free to use it if you’d like.

Thanks again, and congratulations on your excellent work! 🎉
