### 2025-10-17 810b22d
More structure
