+++ docs/contributing.md
@@ -1,0 +1,1 @@
+# Contributing to pyElli
