+++ docs/index.md
@@ -1,4 +1,4 @@
-# Welcome to pyElli
+# Getting Started with pyElli
 
 PyElli is an open-source numerical solver for spectral ellipsometry employing well-known 2x2 and 4x4 solver algorithms. It is intended for a broad range of problems such as simple fitting of layered structures, anisotropic layers and any other polarized light interaction with layered 1D structures. It serves as a system for the day to day ellipsometry task at hand and aims to make optical model generation standardized and reproducible.
 
