+++ docs/howto/custom_fitting.md
@@ -1,0 +1,3 @@
+# Custom Fitting
+
+Custom fitting
