+++ docs/tutorial/simple_fit.md
@@ -1,0 +1,1 @@
+# Performing a Simple Fit
