+++ docs/contributing.md
@@ -1,1 +1,3 @@
-# Contributing to pyElli
+# Contributing
+
+We're happy that you are interested in contributing to pyElli!
