+++ pyproject.toml
@@ -50,13 +50,12 @@ dev = [
     "pytest-cov",
     "nbmake",
     "ruff==0.8.5",
-    "uv",
     "pre-commit",
 ]
 docs = [
     "mkdocs-material",
     "mkdocstrings[python]>=0.18",
-    "mkdocs-autoapi",
+    "mkdocs-api-autonav",
     "mike",
 ]
 
