+++ docs/tutorial/working_with_dispersions.md
@@ -1,1 +1,6 @@
 # Working with Dispersions
+
+TODO:
+- How to construct and add dispersions
+- Formula dispersions
+- How to load dispersions from rii
