+++ docs/reference/kkr.md
@@ -1,0 +1,3 @@
+# Kramers-Kronig Relations
+
+::: elli.kkr.kkr
