+++ pyproject.toml
@@ -56,6 +56,7 @@ docs = [
     "mkdocs-material",
     "mkdocstrings[python]>=0.18",
     "mkdocs-api-autonav",
+    "mkdocs-plotly-plugin",
     "mike",
 ]
 
