### 2025-10-19 127c3fa
add Python Benchmark with pytest-benchmark (pytest) benchmark result for d0e44583dd208a426d65ea42858b5c662d9c78f6
