### 2025-10-22 2663d6e
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 5daac4e3ad0089bc68616bde818cf90d60cfbcb1
