+++ docs/howto/fitting_multiple_angles.md
@@ -1,3 +1,3 @@
-# Custom Fitting
+# Fitting Multiple Angles
 
 Custom fitting
