### 2025-10-27 a6166ab
add Python Benchmark with pytest-benchmark (pytest) benchmark result for de9669bc7b309e2d3bbd7f4a9223328812873749
