### 2025-10-27 0ce841c
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 85e395508b5b26c7f8dec89f4192b40d52ae1bf7
