### 2025-10-27 6831433
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 5352f9937f9e27dfbe06d6113338b05a573caead
