### 2025-10-28 8313bfb
Bump uv from 0.8.6 to 0.9.5 in /docs (#240)
