### 2025-10-28 35acada
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 8313bfb4ebb55825e822b5801795eabf25bd80a4
