### 2025-10-29 db83042
Explain FAIR acronym earlier
