### 2025-10-30 204699a
Bump uv from 0.9.5 to 0.9.6 in /docs (#241)
