### 2025-10-30 22b99a2
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 204699a1b781eca36616d9c40ea448809d8efb84
