+++ docs/index.md
@@ -8,7 +8,7 @@ It serves as a system for the day to day ellipsometry task at hand and aims to m
 PyElli can be installed directly from [PyPI](https://pypi.org/project/pyElli/).
 Installation is as simple as
 
-```
+```sh
 pip install pyelli
 ```
 
@@ -18,6 +18,6 @@ PyElli has the optional dependency `fitting` to install dependencies to use addi
 fitting capabilities and interactive widgets with jupyter notebooks.
 If you want to use those you can directly install the dependencies along with pyElli:
 
-```
+```sh
 pip install 'pyelli[fitting]'
 ```
