+++ docs/tutorial/simple_fit.md
@@ -1,5 +1,27 @@
 # Performing a Simple Fit
 
+In this tutorial we will perform a simple fit with pyElli.
+We determine the thickness of a silicon dioxide layer on top of silicon.
+This fit is often carried out as calibration and benchmark fit for ellipsometers and
+is widely used for quality control of silicon wafers.
+Here, we will use it to make our first steps with pyElli.
+
+Please make sure you meet the following prerequesits:
+
+- You are in a python environment with pyElli [installed](../index.md).
+- We recommend to run the code in a jupyter or jupyter compatible environment.
+However, you _can_ run this code in a repl or python file directly as well.
+It will just not display the fitting widgets.
+- An understanding of ellipsometry, in particular Psi/Delta measurement data.
+
+## Importing the packages
+
+First, we need to import pyElli itself.
+Additionally, we import the `ParamsHist` class, which is a small wrapper around lmfit's `Parameter` class.
+It stores a parameter history to retrieve previous fit configurations.
+We'll also import the function `fit` from pyEllis fitting module.
+This is a function decorator, which we will use to display our interactive fitting widget.
+
 ```python
 import elli
 from elli.fitting import ParamsHist, fit
@@ -7,45 +29,69 @@ from elli.fitting import ParamsHist, fit
 
 ## Reading data
 
+We need some measurement data to perform a fit on and load an example [data set](https://github.com/PyEllips/pyElli/blob/master/examples/Basic%20Usage/SiO2onSi.ellips.nxs) of a psi delta measurement in the standardized [NeXus format](https://manual.nexusformat.org/classes/contributed_definitions/NXellipsometry.html#nxellipsometry) with pyElli's included `read_nexus_psi_delta` function.
+Lets select a measurement angle of 70 degrees (`.loc[ANGLE]`) from the data and
+restrict the wavelength range in between 210 nm and 800 nm (`.loc[210:800]`).
+The restriction is necessary because we're using tabulated literature values for silicon, which are only provided in this wavelength range.
+
 ```python
 ANGLE = 70
-psi_delta = elli.read_nexus_psi_delta("SiO2onSi.ellips.nxs").loc[ANGLE].loc[210:800]
+psi_delta = (
+    elli.read_nexus_psi_delta("SiO2onSi.ellips.nxs")
+    .loc[ANGLE]
+    .loc[210:800]
+)
 ```
 
 ## Setting parameters
 
+To fit our data we need some actual fitting parameters.
+We'll use our `ParamsHist` class which is a wrapper around lmfit's Parameters class.
+The names of the paramters are freely choosable but we recommend to use some material and parameter name combination.
+We'll use standard literature values for the silicon dioxide dispersion parameters $n_0$ (`SiO2_n0`) and $n_1$ (`SiO2_n1`).
+For now, we won't vary these parameters in the fit, hence we set `vary=False` in the parameter definition.
+Last but not least, we'll add a thickness parameter of the silicon dioxide layer with an initial guess of the materials thickness.
+We set the `min` and `max` values to reasonable values and `vary=True` since we want to fit the thickness of the layer.
+
 ```python
 params = ParamsHist()
-params.add("SiO2_n0", value=1.452, min=-100, max=100, vary=False)
-params.add("SiO2_n1", value=36.0, min=-40000, max=40000, vary=False)
-params.add("SiO2_n2", value=0, min=-40000, max=40000, vary=False)
-params.add("SiO2_k0", value=0, min=-100, max=100, vary=False)
-params.add("SiO2_k1", value=0, min=-40000, max=40000, vary=False)
-params.add("SiO2_k2", value=0, min=-40000, max=40000, vary=False)
[…]
