### 2025-11-10 #8503 @zhubonan
Hi @MarJMue I am happy to accept the paper, but before I move this to the next stage, could you please complete the following:
1. Make a software release, please report the version that should be linked with paper here.
2. Archive the release (e.g. your code) with a repository service, e.g. Zenodo and let me know the DOI. Please make sure the title and authors match with that of the paper. Please also make sure that the license matches with that of the software. You can check existing JOSS paper to see how they link with an archive.
3. Please double check the author's ORCID

Thanks!
