### 2025-11-10 a2c539b
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 69163a68ed71957e8261147be50a1451d02cc305
