### 2025-11-10 1841d43
add Python Benchmark with pytest-benchmark (pytest) benchmark result for ab998030549e43a03289c22e4a07abc2f7339159
