### 2025-11-20 #8503 @zhubonan
Hi @MarJMue Please update the title of the Zenodo archive to match that of the paper “pyElli: An open source ellipsometry analysis tool for FAIR data”. You should be able to just edit the metadata without generating a new release.
