+++ docs/tutorial/index.md
@@ -1,1 +1,8 @@
 # Tutorials
+
+Our tutorials provide step-by-step guides for getting started with pyElli.
+We tried to keep everything simple and unambigious, while still providing the necessary details for understanding how things work.
+
+A great first step is to follow the [performing a simple fit tutorial](simple_fit.md).
+The [working with dispersion models](working_with_dispersions.md) covers goes through using pre-built dispersion models,
+load literature values from [refractiveindex.info](https://refractiveindex.info) and explains how to built your own dispersion model.
