### 2025-11-24 84c084b
add Python Benchmark with pytest-benchmark (pytest) benchmark result for b42dac9120e9995bfca4a49d858fa5049c9296b8
