### 2025-12-01 17a4074
Bump fonttools from 4.56.0 to 4.61.0 in /requirements
Bumps [fonttools](https://github.com/fonttools/fonttools) from 4.56.0 to 4.61.0.
- [Release notes](https://github.com/fonttools/fonttools/releases)
- [Changelog](https://github.com/fonttools/fonttools/blob/main/NEWS.rst)
- [Commits](https://github.com/fonttools/fonttools/compare/4.56.0...4.61.0)

---
updated-dependencies:
- dependency-name: fonttools
  dependency-version: 4.61.0
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
