### 2025-12-02 d68a937
Update Citations for tmmax and Byrnes
