+++ paper/paper.bib
@@ -205,6 +205,8 @@
 @misc{byrnes2020multilayer,
   title={Multilayer optical calculations},
   author={Steven J. Byrnes},
+  doi={10.48550/arXiv.1603.02720},
+  url={https://doi.org/10.48550/arXiv.1603.02720},
   year={2020},
   eprint={1603.02720},
   archivePrefix={arXiv},
@@ -417,16 +419,15 @@
   year={2012}
 }
 
-@misc{tmmax,
-  title        = {TMMax: High-performance modeling of multilayer thin-film structures using transfer matrix method with JAX},
-  author       = {Bahrem Serhat Danis and Esra Zayim},
-  year         = {2025},
-  eprint       = {2507.11341},
-  archivePrefix= {arXiv},
-  primaryClass = {physics.comp-ph},
-  doi          = {10.48550/arXiv.2507.11341},
-  url          = {https://doi.org/10.48550/arXiv.2507.11341}
-}
+@article{tmmax, 
+  doi = {10.21105/joss.09088}, 
+  url = {https://doi.org/10.21105/joss.09088}, 
+  year = {2025}, publisher = {The Open Journal}, 
+  volume = {10}, number = {114}, pages = {9088}, 
+  author = {Danis, Bahrem Serhat and Zayim, Esra}, 
+  title = {TMMax: High-performance modeling of multilayer thin-film structures using transfer matrix method with JAX}, 
+  journal = {Journal of Open Source Software} 
+} 
 
 @article{Langevin:24,
   author = {Denis Langevin and Pauline Bennet and Abdourahman Khaireh-Walieh and Peter Wiecha and Olivier Teytaud and Antoine Moreau},
