### 2025-12-02 2550ef6
add Python Benchmark with pytest-benchmark (pytest) benchmark result for f8cbbd443352b7f621ed043ac0a8deec97f8b399
