### 2025-12-02 f685f4d
add Python Benchmark with pytest-benchmark (pytest) benchmark result for b741aa477112a9773263a974cf20c5ef1be4a479
