### 2025-12-06 ccf2b7f
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 3c59cfb1923d7dc725512c1f4616dade83ce32e0
