### 2025-12-06 c882b0a
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 5b0cdbeb7d42798ca8da90c858a4896632b1161a
