### 2025-12-08 381e1e5
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 68b9d4a0fc71b0f5e1819821a30a957a2a716750
