### 2025-12-15 d924d76
add Python Benchmark with pytest-benchmark (pytest) benchmark result for bef5fe29ce41093ce20d0c2d102d6c50fa084020
