### 2025-12-15 c8605e5
add Python Benchmark with pytest-benchmark (pytest) benchmark result for 8e9aa1168a24c4ae22ba8e074f1f187deb921fb2
