### 2025-12-17 984b3e6
add Python Benchmark with pytest-benchmark (pytest) benchmark result for a51038406434f810f0b0a1f2f1290ad17d6741de
